Adding title screen

+++ view/display_score.py
@@ -5,7 +5,7 @@ pygame.font.init()
 score_text_size = 80
 
 score_font = pygame.font.Font(None, score_text_size)
-score_y_placement = int(round(gv.WINDOW_L * 1 / 30))
+score_y_placement = round(gv.WINDOW_L * 1 / 30)
 
 
 def display_score(view, vehicle):
@@ -18,7 +18,7 @@ def display_score(view, vehicle):
     for n in enumerate(score):
         score_w += character_w
 
-    score_x_placement = int(round((gv.WINDOW_W / 2) - (score_w / 2)))
+    score_x_placement = round((gv.WINDOW_W / 2) - (score_w / 2))
 
     """ renders each character spaced out by their individual widths """
     for i, item in enumerate(score):
